sputavaak-nr/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


### dependencies generated by doing
### pipenv requirements | sed 's/.*/"&",/' requirements.txt | grep -iv pypi >tomlreq
### and inserting the contents of tomreq
[project]
name = "sputavaaknr"
version = "0.0.1"
description = "Noise Remover for voice recording CLI"
requires-python = ">=3.10"
dependencies = [
    "appdirs==1.4.4",
    "certifi==2025.8.3; python_version >= '3.7'",
    "cffi==1.17.1; python_version >= '3.8'",
    "charset-normalizer==3.4.3; python_version >= '3.7'",
    "deepfilterlib==0.5.6; python_version >= '3.8'",
    "deepfilternet==0.5.6; python_version >= '3.8' and python_version < '4.0'",
    "ffmpeg-python==0.2.0",
    "filelock==3.19.1; python_version >= '3.9'",
    "fsspec==2025.7.0; python_version >= '3.9'",
    "future==1.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'",
    "idna==3.10; python_version >= '3.6'",
    "jinja2==3.1.6; python_version >= '3.7'",
    "loguru==0.7.3; python_version >= '3.5' and python_version < '4.0'",
    "markupsafe==3.0.2; python_version >= '3.9'",
    "mpmath==1.3.0",
    "networkx==3.5; python_version >= '3.11'",
    "numpy==1.26.4; python_version >= '3.9'",
    "packaging==23.2; python_version >= '3.7'",
    "pycparser==2.22; python_version >= '3.8'",
    "requests==2.32.5; python_version >= '3.9'",
    "setuptools==80.9.0; python_version >= '3.9'",
    "soundfile==0.13.1",
    "sympy==1.14.0; python_version >= '3.9'",
    "torch==2.8.0; python_full_version >= '3.9.0'",
    "torchaudio==2.8.0",
    "typing-extensions==4.14.1; python_version >= '3.9'",
    "urllib3==2.5.0; python_version >= '3.9'",
]

[project.scripts]
sputavaaknr = "sputavaaknr.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests", "examples"]